+++ src/controllers/UserController.ts
@@ -19,10 +19,6 @@ class UserController {
     return this._request(() => this._updatePassword(data), 'update password error');
   }
 
-  // public async getUserById(id: string): Promise<void> {
-  //   return this._request(() => this._getUserById(id), 'get user error');
-  // }
-
   private async _updateProfile(options: UserMainData): Promise<void> {
     const user = await this._api.updateProfile(options);
     store.set('user', user);
@@ -42,7 +38,6 @@ class UserController {
 
   public getUserById(id: number): Promise<User> {
     return this._api.getUserById(id);
-    // store.set('user.current', user);
   }
 
   private _getUsersByLogin(login: string): Promise<User[]> {
