+++ src/modules/BackArrow/BackArrow.ts
@@ -1,6 +1,6 @@
 import ArrowButton from '@/components/ArrowButton';
 import { ArrowButtonPropsType, ArrowButtonType } from '@/components/ArrowButton/ArrowButton';
-import router from '@/utils/components/Router';
+import router from '@/utils/components/Router/Router';
 
 class BackArrow extends ArrowButton<ArrowButtonPropsType> {
   constructor() {
