+++ src/utils/components/BaseComponent/index.ts
@@ -1,0 +1,11 @@
+import BaseComponent from './BaseComponent';
+export type {
+  PropsTypes,
+  ListenersType,
+  ComponentProps,
+  BaseComponentConstructor,
+  IBaseComponent,
+  ComponentDidUpdateType
+} from './BaseComponent';
+
+export default BaseComponent;
