+++ src/utils/components/EventEmitter/index.ts
@@ -1,0 +1,5 @@
+import EventEmitter from './EventEmitter';
+
+export type { CallbackType } from './EventEmitter';
+
+export default EventEmitter;
