+++ src/utils/components/HTTPTransport/index.ts
@@ -1,0 +1,5 @@
+import HTTPTransport from './HTTPTransport';
+
+export { myFetch, fetchWithRetry } from './HTTPTransport';
+
+export default HTTPTransport;
