+++ src/utils/components/Route/Route.ts
@@ -1,9 +1,9 @@
-import renderDOM from '../helpers/renderDOM';
+import renderDOM from '../../helpers/renderDOM';
 import BaseComponent, {
   BaseComponentConstructor,
   IBaseComponent,
   PropsTypes,
-} from './BaseComponent';
+} from '../BaseComponent';
 
 type RouteProps = {
   rootQuery: string;
@@ -35,12 +35,8 @@ class Route<P extends PropsTypes = PropsTypes> {
   }
 
   render() {
-    console.log('render route');
-    console.log('check block', this._block);
     if (!this._block) {
-      console.log('init block');
       this._block = new this._blockClass({ props: {} as P });
-      console.log('block', this._block);
       renderDOM(this._props.rootQuery, this._block as BaseComponent);
       this._block.dispatchComponentDidMount();
       return;
