+++ src/utils/components/Route/index.ts
@@ -1,0 +1,3 @@
+import Route from './Route';
+
+export default Route;
