+++ src/utils/components/Router/index.ts
@@ -1,0 +1,3 @@
+import router from './Router';
+
+export default router;
