+++ src/utils/components/WSTransport/WSTransport.ts
@@ -1,4 +1,4 @@
-import EventEmitter from './EventEmitter';
+import EventEmitter from '../EventEmitter';
 
 export enum SocketEvent {
   MESSAGE = 'message',
