+++ src/utils/components/WSTransport/index.ts
@@ -1,0 +1,5 @@
+import WSTransport from './WSTransport';
+
+export { SocketEvent } from './WSTransport';
+
+export default WSTransport;
